feat: Set DATE_TIME and DATE_TIME_FORMAT in the settings and also env variables
