+++ example.development.env
@@ -35,3 +35,5 @@ tokenLife=7d
 secret=qMyLZALzs229ybdQXNyzYRdju7X784TH
 SPARKPOST_TOKEN=insertokenhere
 TZ=Europe/Stockholm
+DATE_FORMAT=dd-MM-yyyy
+DATE_TIME_FORMAT=dd-MM-yyyy HH:mm
