+++ src/config/ess/configureESSEnvironment.ts
@@ -3,7 +3,7 @@ import { container } from 'tsyringe';
 import { AdminDataSource } from '../../datasources/AdminDataSource';
 import { FeatureId } from '../../models/Feature';
 import { SettingsId } from '../../models/Settings';
-import { setTimezone } from '../setTimezone';
+import { setTimezone, setDateTimeFormats } from '../setTimezoneAndFormat';
 import { Tokens } from '../Tokens';
 
 async function setEssColourTheme() {
@@ -41,4 +41,5 @@ export async function configureESSDevelopmentEnvironment() {
   await setEssColourTheme();
   await enableDefaultEssFeatures();
   await setTimezone();
+  await setDateTimeFormats();
 }
